Added to Tournament Controller

+++ java/src/main/java/com/techelevator/dao/TournamentDAO.java
@@ -22,5 +22,5 @@ public interface TournamentDAO {
 	
 	public void deleteTournament(int id);
 	
-	public void addParticipant(int participantId, int tournamentId);
+	public void addParticipant(long participantId, int tournamentId);
 }
